requirements.txt
Django==5.1.3
django-cors-headers==4.6.0
PyJWT==2.10.0
gunicorn==23.0.0